Menu
Added a menu. So far the Play button only works. The button class is new and it allows you to attach functions to them in runnable variable. Also, the background won't resize for different resolutions or display ratios which needs fixing.
I used GameState enum to switch across these menus.
InputListener uses new hashmap that uses GameState as the key.